+++ TerrariaPack.cs
@@ -104,6 +104,7 @@ public sealed class Terraria : SimpleTCPPack
                 new Effect("- Confuse", "buff_confuse", BuffFolder) {Price = 30, Description = "Temporarily reverse the streamer's controls"},
                 new Effect("- Make invisible", "buff_invisible", BuffFolder) {Price = 30, Description = "Temporarily make the streamer invisible"},
                 new Effect("- Blind", "buff_blind", BuffFolder) {Price = 30, Description = "Temporarily decrease the streamer's vision"},
+                new Effect("- Curse", "buff_curse", BuffFolder) {Price = 20, Description = "Temporarily prevent the streamer from using any items"},
 
                 // --- Inventory effects
                 new Effect("Inventory effects", "inventory_folder", ItemKind.Folder),